@@ -11,7 +11,7 @@ using WebDatabasesMVC.Models;
 
 namespace WebDatabasesMVC.Controllers
 {
-    public class HomeController : Controller
+    public class HomeController : Microsoft.AspNetCore.Mvc.Controller
     {
         private readonly ILogger<HomeController> _logger;
 
