@@ -120,13 +120,13 @@ namespace WebDatabasesMVC.Controllers
 
 
         // GET: TestPageController/Details/5
-        public ActionResult Details(int id)
+        public Microsoft.AspNetCore.Mvc.ActionResult Details(int id)
         {
             return View();
         }
 
         // GET: TestPageController/Create
-        public ActionResult Create()
+        public Microsoft.AspNetCore.Mvc.ActionResult Create()
         {
             return View();
         }
@@ -134,7 +134,7 @@ namespace WebDatabasesMVC.Controllers
         // POST: TestPageController/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create(IFormCollection collection)
+        public Microsoft.AspNetCore.Mvc.ActionResult Create(IFormCollection collection)
         {
             try
             {
@@ -147,7 +147,7 @@ namespace WebDatabasesMVC.Controllers
         }
 
         // GET: TestPageController/Edit/5
-        public ActionResult Edit(int id)
+        public Microsoft.AspNetCore.Mvc.ActionResult Edit(int id)
         {
             return View();
         }
@@ -155,7 +155,7 @@ namespace WebDatabasesMVC.Controllers
         // POST: TestPageController/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(int id, IFormCollection collection)
+        public Microsoft.AspNetCore.Mvc.ActionResult Edit(int id, IFormCollection collection)
         {
             try
             {
@@ -168,7 +168,7 @@ namespace WebDatabasesMVC.Controllers
         }
 
         // GET: TestPageController/Delete/5
-        public ActionResult Delete(int id)
+        public Microsoft.AspNetCore.Mvc.ActionResult Delete(int id)
         {
             return View();
         }
@@ -176,7 +176,7 @@ namespace WebDatabasesMVC.Controllers
         // POST: TestPageController/Delete/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Delete(int id, IFormCollection collection)
+        public Microsoft.AspNetCore.Mvc.ActionResult Delete(int id, IFormCollection collection)
         {
             try
             {
