@@ -1,4 +1,4 @@
-﻿@model BusinessLibrary.BusinessClasses.WebDbFrequencyCrosstabAndMapSelectionCriteria
+@model BusinessLibrary.BusinessClasses.WebDbFrequencyCrosstabAndMapSelectionCriteria
 @using BusinessLibrary.BusinessClasses
 @using Microsoft.AspNetCore.Html;
 
@@ -11,7 +11,7 @@
         <div class="card">
             <div class="card-body">
                 <br />
-                <iframe src="@(new HtmlString(@Model.nameXAxis))" style="width:100%;height:800px;" title="eppiMapperMap"></iframe>                  
+@{ Microsoft.AspNetCore.Html. }                <iframe src="@(new HtmlString(@Model.nameXAxis))" style="width:100%;height:800px;" title="eppiMapperMap"></iframe>                  
             </div>
         </div>
     </div>
