@@ -1,4 +1,4 @@
-﻿
+
 @using System.IO;
 @using System.Security.Claims;
 @using Microsoft.AspNetCore.Html;
@@ -120,7 +120,7 @@
                 {
                     if (imgUrl3 != "")
                     {
-                        <a href="@(new HtmlString(imgUrl3))" class="logo" target="_blank" title="Open this Organisation Webpage">
+@{ Microsoft.AspNetCore.Html. }                        <a href="@(new HtmlString(imgUrl3))" class="logo" target="_blank" title="Open this Organisation Webpage">
                             <img src="@imgPath3" alt="logo image" class="navbar-brand" style="max-width: 80px; max-height: 55px; object-fit: cover;">
                         </a>
                     }
