@@ -1,4 +1,4 @@
-﻿<!-- Review -->
+<!-- Review -->
 
 @model WebDbWithRevInfo;
 @using WebDatabasesMVC.ViewModels;
@@ -568,7 +568,7 @@
                                 </div>
                                 <div class="card-body">
 
-                                    <div class="card-category">@(new HtmlString(Model.WebDb.WebDBDescription))</div>
+@{ Microsoft.AspNetCore.Html. }                                    <div class="card-category">@(new HtmlString(Model.WebDb.WebDBDescription))</div>
 
 
 
